move AzureVMMetadata to models folder
